+++ CHANGELOG.md
@@ -20,3 +20,4 @@ Changes:
 Changes:
 * Added `.max_size()` function
 * `.get()` and `.get_exact()` now return a reference to the `BumpyEntry` rather than a new `BumpyEntry` with a reference to the object `T`
+* Add `.get_mut()` and `.get_exact_mut()` functions